More bugfixes
The order of the admin menu has changed according to the work process.
User section of the admin functions has been implemented. Deletion of
Bilfabrikant is stopped if it contain active carmodels.

+++ includes/admin/delete_manufacturer.php
@@ -2,7 +2,18 @@
 	$m_id = $_POST['manufacturer_id']; // Kun tabellens ID behoeves for at identificere og slette posten
 	
 	if ($m_id!=NULL) {
-		mysql_query("DELETE FROM fabrikant WHERE fabrikant_id = '$m_id' LIMIT 1;") 
+
+		$res_check = mysql_query("SELECT * FROM biler WHERE fabrikant_id = '$m_id'", $conn); 
+		$num_rows = mysql_num_rows($res_check);
+		
+		echo $num_rows;
+		
+		if ($num_rows==NULL) {
+			mysql_query("DELETE FROM fabrikant WHERE fabrikant_id = '$m_id' LIMIT 1;") 
 			or die(mysql_error());
+		} else {
+			echo "<p>Fejl: Du kan ikke slette bilfabrikant, når der er biler tilknyttet</p>";
+			break;
+		}
 	}
 ?>
